Refactor tailscope-core lib into focused modules
